I reorganized the repo to look more clean
I also updated the Database with some more data.
