+++ README.md
@@ -1,3 +1,3 @@
-# govhack2020
+# GovHack2020
 
-This is the project for GovHack2020.
+Our project uses IoT devices to track plastic waste. <We can add more detail to this, then create a github pages>
