--- m_app.html
@@ -1,7 +1,0 @@
-<form method="post">
-        <input type="text" name="date" class="form-control"></input>
-        <input type="text" name="time" class="form-control"></input>
-        <input type="text" name="type" class="form-control"></input>
-        <input type="text" name="lon" class="form-control"></input>
-        <input type="text" name="lat" class="form-control"></input>
-</form>
